preprocess: move emsc and integrate test in separate files
